Allow public and private releases

+++ ci/upload_conda_pkg.sh
@@ -4,5 +4,13 @@
 # secrets (DOCKER_PVTKEY and MORIA_USERNAME) in the log of the workflow files.
 # All secrets should be passed as envvars
 
-echo "${DOCKER_PVTKEY}"  > ~/.ssh/id_rsa.pub && chmod 400 ~/.ssh/id_rsa
-rsync -av --delete ./conda-pkgs/ "${MORIA_USERNAME}@moria.chem.ufl.edu:/data/conda-pkgs/"
+if [ "$1" = internal ]; then
+    echo "${DOCKER_PVTKEY}"  > ~/.ssh/id_rsa.pub && chmod 400 ~/.ssh/id_rsa
+    rsync -av --delete ./conda-pkgs/ "${MORIA_USERNAME}@moria.chem.ufl.edu:/data/conda-pkgs/"
+elif [ "$1" = public ]; then
+    anaconda \
+        --token "${CONDA_TOKEN}" \
+        upload \
+            --user roitberg-group \
+            --force ./conda-pkgs/linux-64/*.tar.gz
+fi
